% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ef_emep.R
\name{ef_emep}
\alias{ef_emep}
\title{Emissions factors from European European Environment Agency}
\usage{
ef_emep(nfr = "1.A.1.a", tier, fuel, abatement, pol, returnfdb = FALSE)
}
\arguments{
\item{nfr}{String: nomenclature for reporting}

\item{tier}{numeric 1 or 2}

\item{fuel}{String for type of fuel.}

\item{abatement}{String.}

\item{pol}{String for pollutant.}

\item{returnfdb}{logical return directly db}
}
\value{
Return a data.table
}
\description{
Emissions factors from European European Environment Agency
}
\examples{
{
# ef_eea(category = "NAO SEI")
}
}
\keyword{emission}
\keyword{factors}
