% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/esri_to_sf_geom.R
\name{esri_to_sf_geom}
\alias{esri_to_sf_geom}
\title{get features}
\usage{
esri_to_sf_geom(json_feats, geom_type, crs = 4326)
}
\arguments{
\item{json_feats}{Character}

\item{geom_type}{Character.}

\item{crs}{CRS.}
}
\description{
get features.
}
\examples{
\dontrun{
# do not run
}
}
\seealso{
\code{\link{layer_info}}
}
