% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/esri_to_sf_point.R
\name{esri_to_sf_point}
\alias{esri_to_sf_point}
\title{get features}
\usage{
esri_to_sf_point(features)
}
\arguments{
\item{features}{Character}
}
\description{
get features.
}
\examples{
\dontrun{
# do not run
}
}
\seealso{
\code{\link{layer_info}}
}
