% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_esri_features.R
\name{get_esri_features}
\alias{get_esri_features}
\title{get features}
\usage{
get_esri_features(
  url = paste0("http://portal1.snirh.gov.br/ana/",
    "rest/services/Esta\%C3\%A7\%C3\%B5es_da_",
    "Rede_Hidrometeorol\%C3\%B3gica_Nacional", "_em_Opera\%C3\%A7\%C3\%A3o/MapServer/1"),
  query_url = paste(url, "query", sep = "/"),
  fields = c("*"),
  where = "1=1",
  token = ""
)
}
\arguments{
\item{url}{Character, default .}

\item{query_url}{Character.}

\item{where}{Character.}

\item{token}{Character.}

\item{ssl}{Logical, default = FALSE.}
}
\description{
get features.
}
\examples{
\dontrun{
# do not run
(geom_type <- layer_info()$geometryType)
# rede
url1 = paste0("http://portal1.snirh.gov.br/ana/",
"rest/services/Esta\%C3\%A7\%C3\%B5es_da_",
"Rede_Hidrometeorol\%C3\%B3gica_Nacional",
"_em_Opera\%C3\%A7\%C3\%A3o/MapServer/1")
# hidroweb
url1 = paste0("http://portal1.snirh.gov.br/ana/",
"rest/services/Esta\%C3\%A7\%C3\%B5es_da_",
"Rede_Hidrometeorol\%C3\%B3gica_Nacional",
"_em_Opera\%C3\%A7\%C3\%A3o/MapServer/4")
esri_features <- get_esri_features(url2)
simple_features <- esri_to_sf_geom(esri_features, geom_type = geom_type)
}
}
\seealso{
\code{\link{layer_info}}
}
