% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_esri_features_by_ids.R
\name{get_esri_features_by_ids}
\alias{get_esri_features_by_ids}
\title{get features}
\usage{
get_esri_features_by_ids(
  ids,
  url = paste0("http://portal1.snirh.gov.br/ana/",
    "rest/services/Esta\%C3\%A7\%C3\%B5es_da_",
    "Rede_Hidrometeorol\%C3\%B3gica_Nacional", "_em_Opera\%C3\%A7\%C3\%A3o/MapServer/1"),
  query_url = paste(url, "query", sep = "/"),
  fields = c("*"),
  token = "",
  ssl = FALSE,
  simplifyDataFrame = FALSE,
  simplifyVector = FALSE,
  full = FALSE
)
}
\arguments{
\item{ids}{Integer.}

\item{url}{Character.}

\item{query_url}{Character.}

\item{fields}{Character.}

\item{token}{Character.}

\item{ssl}{Logical, default = FALSE.}

\item{simplifyDataFrame}{Logical, default = FALSE.}

\item{simplifyVector}{Logical, default = FALSE.}

\item{full}{Logical, default = FALSE.}
}
\description{
get features.
}
\examples{
\dontrun{
# do not run
ids <- get_object_ids()
feat <- get_esri_features_by_ids(ids = ids)
}
}
\seealso{
\code{\link{layer_info}}
}
