% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layer_info.R
\name{layer_info}
\alias{layer_info}
\title{get layer info}
\usage{
layer_info(
  url = paste0("http://portal1.snirh.gov.br/ana/",
    "rest/services/Esta\%C3\%A7\%C3\%B5es_da_",
    "Rede_Hidrometeorol\%C3\%B3gica_Nacional", "_em_Opera\%C3\%A7\%C3\%A3o/MapServer/1"),
  token = "",
  ssl = FALSE
)
}
\arguments{
\item{url}{Character, default .}

\item{token}{Character.}

\item{ssl}{Logical, default = FALSE.}
}
\description{
get layer info.
}
\examples{
\dontrun{
# do not run
a <- layer_info()
names(a)
a$geometryType
}
}
\seealso{
\code{\link{layer_info}}
}
