% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pega-package.R
\docType{package}
\name{pega-package}
\alias{pega}
\alias{pega-package}
\title{pega: Pollutants Emissions and Gaseous of the Atmosphere}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Elaboration of fix emissions inventories, consisting in four stages, pre-processing activity data, preparing emissions factors, estimating the emissions and post-processing of emissions in maps and databases. More details in \doi{}. Before using PEGA you need to know the industrial composition of your study area, in other words, the combination of industrial type, size and fuel of the fleet. Then, it is recommended to start with the project to download a template to create a structure of directories and scripts.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/atmoschem/pega}
  \item Report bugs at \url{https://github.com/atmoschem/pega/issues}
}

}
\author{
\strong{Maintainer}: Tailine Santos \email{tailine.atm@gmail.com} (\href{https://orcid.org/0000-0001-7723-9050}{ORCID})

Authors:
\itemize{
  \item Sergio Ibarra-Espinosa \email{zergioibarra@gmail.com} (\href{https://orcid.org/0000-0002-3162-1905}{ORCID})
}

Other contributors:
\itemize{
  \item Iara Silva \email{iaradasilva@usp.br} (\href{https://orcid.org/0000-0003-1306-5906}{ORCID}) [contributor]
}

}
\keyword{internal}
